Styling
I've near enough finished the styling for how I want the application to look (on mobile). I may tweak a few things here and there.

I've used 'font-awesome' for the icons which has proven to be a fantastic resource. This has been styled for mobile screens initially, I'll need a media query for wider screens. I haven't done this yet.