Added limit
In this commit I've added a limit on the number of to-dos that can be added. When the limit has been reached, a further addition will be met with a pop up message.

I've also linked the media quieries to the to-do limit, changing it depending on the screen size. The bigger the screen, the more to-dos can be added.